Fix:UserRepository id 타입 선언 잘못 되어있던것 변경!

+++ RecipiaProject/src/main/java/com/ll/exam/RecipiaProject/mypage/MyPageService.java
@@ -12,7 +12,7 @@ public class MyPageService {
     private final UserRepository userRepository;
 
     //id 가져오기
-    public SiteUser getUser(Long id){
+    public SiteUser getUser(Integer id){
         return userRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("no %d user not found,".formatted(id)));
     }
